requirements.txt
django
channels
locked_dict
opencv-python
Pillow
scikit-image
tqdm
numpy
python-bidi
edit_distance
xlsxwriter
lxml
protobuf
